+++ docs/components/content/Alert/AlertClose.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+  import { Alert } from '@lite-space/ui'
+</script>
+
+<template>
+  <Alert
+    icon="i-heroicons:at-symbol-16-solid"
+    message="Heads up!"
+    description="You can add components to your app using the cli."
+    closable
+  />
+</template>