+++ packages/components/Alert/src/types.d.ts
@@ -1,0 +1,9 @@
+interface LtAlertProps {
+  icon: string
+  message: string
+  description: string
+  closable: boolean
+}
+
+export type AlertProps = Partial<LtAlertProps>
+export type AlertInstance = InstanceType<typeof import('./Alert.vue')['default']>