+++ packages/theme/src/shortcuts/Alert.ts
@@ -1,0 +1,9 @@
+export const Alert = {
+  'lt-alert': 'flex justify-between items-start',
+  'lt-alert-inner': 'flex items-start gap-2',
+  'lt-alert-icon': 'inline-block align-middle text-base',
+  'lt-alert-message': '',
+  'lt-alert-description': '',
+  'lt-alert-closeIcon': 'inline-block text-lg cursor-pointer transition text-gray-500 hover:text-gray-700'
+
+}