+++ packages/theme/src/shortcuts/Card.ts
@@ -1,7 +1,8 @@
 export const Card = {
   'lt-card': 'border border-solid border-gray-200 rounded-md',
-  'lt-card-header': 'flex text-sm items-center px-4 py-3 border-b border-b-solid border-b-gray-200',
-  'lt-card-header-title': 'flex-1 font-600 text-base',
-  'lt-card-body': 'p-4',
-  'lt-card-footer': 'px-4 py-3 text-sm border-t border-t-solid border-t-gray-200'
+  'lt-card-header': 'flex text-sm items-center p-3 py-2 border-b border-b-solid border-b-gray-200',
+  'lt-card-header-title': 'flex-1 font-600 text-sm',
+  'lt-card-body': 'px-3 py-2 text-sm',
+  'lt-card-extra': 'text-sm',
+  'lt-card-footer': 'px-3 py-2 text-sm border-t border-t-solid border-t-gray-200'
 }