+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lite-space/theme
 
+## 0.5.0
+
+### Patch Changes
+
+- @lite-space/alert@0.0.0
+  @lite-space/card@0.0.4
+  @lite-space/ui@0.5.0
+  @lite-space/theme@0.5.0
+
 ## 0.4.2
 
 ### Patch Changes